requirements.txt
astroid==2.12.12
attrs==22.1.0
black==22.10.0
cfgv==3.3.1
click==8.1.3
colorama==0.4.5
dill==0.3.5.1
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
greenlet==1.1.3.post0
identify==2.5.6
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.20.0
py==1.11.0
pylint==2.15.4
pyparsing==3.0.9
pytest==7.1.3
PyYAML==6.0
SQLAlchemy==1.4.42
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
validator.py==1.3.0
virtualenv==20.16.5
Werkzeug==2.2.2
wrapt==1.14.1
